refactor: standardize chai imports and fix module paths across test files
