+++ test/adb/command/host-transport/ipRules.ts
@@ -1,7 +1,7 @@
-import Chai, { expect } from 'chai';
+import { expect, use } from 'chai';
 import simonChai from 'sinon-chai';
-Chai.use(simonChai);
-import { IpRuleCommand } from '../../../../src/adb/command/host-transport.js';
+use(simonChai);
+import IpRuleCommand from '../../../../src/adb/command/host-transport/ipRule.js';
 import Tester from './Tester.js';
 
 const t = new Tester(IpRuleCommand);
