+++ test/adb/command/host-transport/local.ts
@@ -1,7 +1,7 @@
 import Stream from 'node:stream';
-import Chai, { expect } from 'chai';
+import { expect, use } from 'chai';
 import simonChai from 'sinon-chai';
-Chai.use(simonChai);
+use(simonChai);
 import MockConnection from '../../../mock/connection.js';
 import Protocol from '../../../../src/adb/protocol.js';
 import LocalCommand from '../../../../src/adb/command/host-transport/local.js';
