+++ test/adb/command/host-transport/monkey.ts
@@ -1,8 +1,8 @@
 import Stream from 'node:stream';
 import { setImmediate } from "node:timers";
-import Chai, { expect } from 'chai';
+import { expect, use } from 'chai';
 import simonChai from 'sinon-chai';
-Chai.use(simonChai);
+use(simonChai);
 import MockConnection from '../../../mock/connection.js';
 import Protocol from '../../../../src/adb/protocol.js';
 import MonkeyCommand from '../../../../src/adb/command/host-transport/monkey.js';
