+++ test/adb/command/host-transport/shell.ts
@@ -1,8 +1,7 @@
 import { setImmediate } from "node:timers";
-
-import Chai, { expect } from 'chai';
+import { expect, use } from 'chai';
 import simonChai from 'sinon-chai';
-Chai.use(simonChai);
+use(simonChai);
 import MockConnection from '../../../mock/connection.js';
 import Protocol from '../../../../src/adb/protocol.js';
 import { AdbFailError } from '../../../../src/index.js';
