+++ test/adb/command/host-transport/tcpip.ts
@@ -1,6 +1,6 @@
-import Chai, { expect } from 'chai';
+import { expect, use } from 'chai';
 import simonChai from 'sinon-chai';
-Chai.use(simonChai);
+use(simonChai);
 import MockConnection from '../../../mock/connection.js';
 import Protocol from '../../../../src/adb/protocol.js';
 import TcpIpCommand from '../../../../src/adb/command/host-transport/tcpip.js';
