+++ test/adb/command/host/disconnect.ts
@@ -1,7 +1,7 @@
 import { setImmediate } from "node:timers";
-import Chai, { expect } from 'chai';
+import {use, expect } from 'chai';
 import simonChai from 'sinon-chai';
-Chai.use(simonChai);
+use(simonChai);
 import MockConnection from '../../../mock/connection.js';
 import Protocol from '../../../../src/adb/protocol.js';
 import HostDisconnectCommand from '../../../../src/adb/command/host/HostDisconnectCommand.js';
