+++ test/adb/linetransform.ts
@@ -1,7 +1,7 @@
 import { Buffer } from 'node:buffer';
 import Stream from 'node:stream';
 import simonChai from 'sinon-chai';
-import { use, expect } from 'chai';
+import { expect, use } from 'chai';
 import LineTransform from '../../src/adb/linetransform.js';
 import MockDuplex from '../mock/duplex.js';
 use(simonChai);
